# TagHelperDescriptor Hashing and Caching