--- a/src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/MemoryCache.cs
+++ b/src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/MemoryCache.cs
@@ -48,6 +48,11 @@ namespace Microsoft.CodeAnalysis.Razor
                 Compact();
             }
 
+            if (_dict.ContainsKey(key))
+            {
+                _dict.Remove(key);
+            }
+
             _dict.Add(key, new CacheEntry
             {
                 LastAccess = DateTime.UtcNow,