--- a/src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/Serialization/JsonReaderExtensions.cs
+++ b/src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/Serialization/JsonReaderExtensions.cs
@@ -2,7 +2,6 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.Collections.Generic;
 using System.Diagnostics;
 using Newtonsoft.Json;
 
@@ -53,6 +52,69 @@ namespace Microsoft.CodeAnalysis.Razor.Serialization
             }
 
             throw new JsonSerializationException($"Could not find string property '{propertyName}'.");
+        }
+
+        public static (int? hash, string propertyValue) ReadPropertyAndHash(this JsonReader reader, string propertyName)
+        {
+            int? hash = null;
+
+            while (reader.Read())
+            {
+                switch (reader.TokenType)
+                {
+                    case JsonToken.PropertyName:
+                        var curProperty = reader.Value.ToString();
+                        if (reader.Read())
+                        {
+                            if (curProperty == RazorSerializationConstants.HashCodePropertyName)
+                            {
+                                hash = Convert.ToInt32(reader.Value);
+                            }
+                            else if (curProperty == propertyName)
+                            {
+                                var value = (string)reader.Value;
+                                return (hash, value);
+                            }
+                            else
+                            {
+                                throw new JsonSerializationException($"Encountered unknown property when looking for hash or '{propertyName}'.");
+                            }
+                        }
+                        else
+                        {
+                            return default;
+                        }
+                        break;
+                }
+            }
+
+            throw new JsonSerializationException($"Could not find string property '{propertyName}' with hash.");
+        }
+
+        public static void ReadToEndOfCurrentObject(this JsonReader reader)
+        {
+            var nestingLevel = 0;
+
+            while (reader.Read())
+            {
+                switch (reader.TokenType)
+                {
+                    case JsonToken.StartObject:
+                        nestingLevel++;
+                        break;
+                    case JsonToken.EndObject:
+                        nestingLevel--;
+
+                        if (nestingLevel == -1)
+                        {
+                            return;
+                        }
+
+                        break;
+                }
+            }
+
+            throw new JsonSerializationException($"Could not read till end of object, end of stream. Got '{reader.TokenType}'.");
         }
     }
 }