--- a/src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/Serialization/RazorSerializationConstants.cs
+++ b/src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/Serialization/RazorSerializationConstants.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace Microsoft.CodeAnalysis.Razor.Serialization
+{
+    internal static class RazorSerializationConstants
+    {
+        public static readonly string HashCodePropertyName = "HashCode";
+    }
+}