--- a/src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/TagHelperDescriptorCache.cs
+++ b/src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/TagHelperDescriptorCache.cs
@@ -10,10 +10,10 @@ namespace Microsoft.CodeAnalysis.Razor
         private static readonly MemoryCache<int, TagHelperDescriptor> CachedTagHelperDescriptors =
             new MemoryCache<int, TagHelperDescriptor>(1500);
 
-        public static bool TryGetDescriptor(int hashCode, out TagHelperDescriptor descriptor) =>
+        internal static bool TryGetDescriptor(int hashCode, out TagHelperDescriptor descriptor) =>
             CachedTagHelperDescriptors.TryGetValue(hashCode, out descriptor);
 
-        public static void Set(int hashCode, TagHelperDescriptor descriptor) =>
+        internal static void Set(int hashCode, TagHelperDescriptor descriptor) =>
             CachedTagHelperDescriptors.Set(hashCode, descriptor);
     }
 }