2/2 PRs for https://github.com/dotnet/aspnetcore/issues/23170

Other PR: https://github.com/dotnet/aspnetcore/pull/24551

- Used the project for a short period of time
	- Created a file, navigated between files, made some edits.
	- I ran into 238 cache misses (unique taghelper allocations) and 4612 cache hits
	- 95%+ cache hit
	- Scales up incredibly well. Number of TagHelperDescriptors unlikely to change unless
		- New files added
		- New assemblies / packages added
	- Consequently, each subsequent `GetTagHelpers` call through Roslyn is going to save ~230 tag helper allocations as we'll be able to leverage cached values.

### Benchmarks
|                   Method |     Mean |    Error |   StdDev |    Op/s | Allocated |
|------------------------- |---------:|---------:|---------:|--------:|----------:|
| 'TagHelpers GetHashCode' | 0.137 ms | 1.279 us | 1.196 us | 7,306.8 |  17.93 KB |

The following benchmark demonstrates cached/uncached performance. I believe these benchmarks aren't showing the full picture due to how BenchmarkDotNet deals with micro-benchmarks and repeated runs. From the above experience, this should perform even better in practice. 

|                                    Method          |     Mean |     Error |    StdDev |  Op/s |   Gen 0 |   Gen 1 |   Gen 2 | Allocated |
|----------------------------------------------------|---------:|----------:|----------:|------:|--------:|--------:|--------:|----------:|
| 'Cached Razor TagHelper Roundtrip Serialization'   | 3.243 ms | 0.0634 ms | 0.0593 ms | 308.4 | 42.9688 | 39.0625 | 35.1563 |   1.95 MB |
| 'Uncached Razor TagHelper Roundtrip Serialization' | 4.891 ms | 0.1082 ms | 0.0959 ms | 204.4 | 78.1250 | 54.6875 | 39.0625 |   3.85 MB |

 Note `GetHashCode` has minimal perf overhead in comparison to the underlying serialization/deserialization cost. As a result we get 33% faster roundtrip serialization and half the memory utilization.

|                                    Method |     Mean |     Error |    StdDev |  Op/s |   Gen 0 |   Gen 1 |   Gen 2 | Allocated |
|------------------------------------------ |---------:|----------:|----------:|------:|--------:|--------:|--------:|----------:|
|    'Before Razor TagHelper Serialization' | 1.782 ms | 0.0353 ms | 0.0378 ms | 561.3 | 35.1563 | 35.1563 | 35.1563 |   1.11 MB |
|     'After Razor TagHelper Serialization' | 1.932 ms | 0.0385 ms | 0.0632 ms | 517.6 | 35.1563 | 35.1563 | 35.1563 |   1.13 MB |
|  'Before Razor TagHelper Deserialization' | 4.198 ms | 0.0173 ms | 0.0161 ms | 238.2 | 39.0625 | 15.6250 |       - |   3.28 MB |
|   'After Razor TagHelper Deserialization' | 4.279 ms | 0.0215 ms | 0.0201 ms | 233.7 | 39.0625 | 15.6250 |       - |   3.29 MB |

Note, marginally higher CPU/Memory for unidirectional serialization/deserialization due to cost of hashing. Note this benchmark would not benefit for caching.