@NTaylorMullen 2020-08-04
I'm  actually  quite surprised  by  the  "Cached  Razor TagHelper Roundtrip Serialization"  only being a 33%  increase. I  would  have  thought it'd  be  significantly more. Does that meant he majority  of the compute time is  at the  JsonReader  layer?