@TanayParikh 2020-08-05
> I'm actually quite surprised by the "Cached Razor TagHelper Roundtrip Serialization" only being a 33% increase. I would have thought it'd be significantly more. Does that meant he majority of the compute time is at the JsonReader layer?

Yeah I expected the perf impact to be even more significant as well. I don't think these benchmarks are doing these changes justice.

Some more data points:
|                                                 Method |     Mean |    Error |   StdDev |  Op/s |     Gen 0 |     Gen 1 |    Gen 2 | Allocated |
|--------------------------------------------------------|---------:|---------:|---------:|------:|----------:|----------:|---------:|----------:|
| 'Uncached 25x Razor TagHelper Roundtrip Serialization' | 121.7 ms | 2.423 ms | 2.148 ms | 8.219 | 1875.0000 | 1750.0000 | 937.5000 |  96.19 MB |
|   'Cached 25x Razor TagHelper Roundtrip Serialization' | 85.59 ms | 1.390 ms | 1.300 ms | 11.68 | 1250.0000 |  937.5000 | 937.5000 |  48.69 MB |

Same roundtrip benchmarks as before but looped 25x - seeing if benchmarkdotnet isn't allowing for appropriate caching. It seems to be fine.

> Does that meant he majority of the compute time is at the JsonReader layer?

That appears to be the case.

"Round trip with caching":
|                                    Method |       Mean |      Error |     StdDev |    Op/s |   Gen 0 |   Gen 1 |   Gen 2 |  Allocated |
|------------------------------------------ |-----------:|-----------:|-----------:|--------:|--------:|--------:|--------:|-----------:|
| 'Razor TagHelper Roundtrip Serialization' | 3,426.7 us | 59.2265 us | 55.4005 us |   291.8 | 46.8750 | 35.1563 | 35.1563 | 1994.55 KB |
|         'Razor TagHelper Deserialization' | 4,271.1 us | 28.0137 us | 21.8713 us |   234.1 | 39.0625 | 15.6250 |       - | 3365.76 KB |

"Round trip with deserialization just being `ReadToEndOfCurrentObject`" (aka just the JsonReader overhead):
|                                    Method |       Mean |      Error |     StdDev |    Op/s |   Gen 0 |   Gen 1 |   Gen 2 |  Allocated |
|------------------------------------------ |-----------:|-----------:|-----------:|--------:|--------:|--------:|--------:|-----------:|
| 'Razor TagHelper Roundtrip Serialization' | 3,372.0 us | 65.6814 us | 61.4384 us |   296.6 | 46.8750 | 35.1563 | 35.1563 | 1994.54 KB |
|         'Razor TagHelper Deserialization' | 2,269.2 us | 13.7223 us | 12.8358 us |   440.7 | 11.7188 |       - |       - | 1070.68 KB |

From `Razor TagHelper Deserialization` we see the `JsonReader` overhead for serialization is roughly half the total time expected for deserialization. So the overhead is substantial. Note the `Razor TagHelper Deserialization` uses taghelpers without hashed values.

The `Razor TagHelper Roundtrip Serialization` differs by ~50us (1.5%) between retrieving from cache/deserializing vs. doing nothing at all (just reading till end of object). This strongly indicates the `JsonReader` is the remaining bottleneck. 

A potential "solution" for this is for us to simply not read. This would corrupt the `JsonReader` state and would definitely be a bad practice. Probably best to leave as is.