@TanayParikh 2020-08-05
With current read to end of object
|                                    Method |     Mean |     Error |    StdDev |  Op/s |   Gen 0 |   Gen 1 |   Gen 2 | Allocated |
|------------------------------------------ |---------:|----------:|----------:|------:|--------:|--------:|--------:|----------:|
| 'Razor TagHelper Roundtrip Serialization' | 3.419 ms | 0.0659 ms | 0.0810 ms | 292.5 | 42.9688 | 31.2500 | 31.2500 |   1.95 MB |
|         'Razor TagHelper Deserialization' | 2.319 ms | 0.0385 ms | 0.0361 ms | 431.2 | 11.7188 |       - |       - |   1.05 MB |


With `[JsonReader.Skip](https://www.newtonsoft.com/json/help/html/Methods_T_Newtonsoft_Json_JsonReader.htm)`:
|                                    Method |     Mean |     Error |    StdDev |  Op/s |   Gen 0 |   Gen 1 |   Gen 2 | Allocated |
|------------------------------------------ |---------:|----------:|----------:|------:|--------:|--------:|--------:|----------:|
| 'Razor TagHelper Roundtrip Serialization' | 3.365 ms | 0.0551 ms | 0.0516 ms | 297.2 | 46.8750 | 35.1563 | 35.1563 |   1.95 MB |
|         'Razor TagHelper Deserialization' | 2.371 ms | 0.0075 ms | 0.0067 ms | 421.8 | 11.7188 |       - |       - |   1.05 MB |