@TanayParikh 2020-08-05
Moving the hashing/caching up to the `TagHelperResolutionResult` layer without reading through the result if we're able to find a cached value:

1x
|                                                                           Method |     Mean |     Error |    StdDev |  Op/s |   Gen 0 |   Gen 1 |   Gen 2 | Allocated |
|------------------------------------------------------------------------          |---------:|----------:|----------:|------:|--------:|--------:|--------:|----------:|
| 'Cached Razor TagHelperResolutionResult Roundtrip JsonConverter Serialization'   | 1.844 ms | 0.0339 ms | 0.0317 ms | 542.3 | 41.0156 | 37.1094 | 37.1094 |   1.21 MB |
| 'Uncached Razor TagHelperResolutionResult Roundtrip JsonConverter Serialization' | 3.652 ms | 0.0658 ms | 0.0615 ms | 273.8 | 42.9688 | 35.1563 | 31.2500 |   2.04 MB |
 
25x
|                                                      Method |     Mean |     Error |   StdDev |  Op/s |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|------------------------------------------------------------ |---------:|----------:|---------:|------:|---------:|---------:|---------:|----------:|
|'Cached TagHelperResolutionResult Roundtrip Serialization'   | 45.94 ms | 0.9160 ms | 1.555 ms | 21.77 | 937.5000 | 875.0000 | 875.0000 |  30.24 MB |
|'Uncached TagHelperResolutionResult Roundtrip Serialization' | 91.45 ms |  1.236 ms | 1.156 ms | 10.93 | 1125.0000| 937.5000 | 812.5000 |  51.01 MB |

vs. (from above)
Some more data points:
|                                                 Method |     Mean |    Error |   StdDev |  Op/s |     Gen 0 |     Gen 1 |    Gen 2 | Allocated |
|--------------------------------------------------------|---------:|---------:|---------:|------:|----------:|----------:|---------:|----------:|
| 'Uncached 25x Razor TagHelper Roundtrip Serialization' | 121.7 ms | 2.423 ms | 2.148 ms | 8.219 | 1875.0000 | 1750.0000 | 937.5000 |  96.19 MB |
|   'Cached 25x Razor TagHelper Roundtrip Serialization' | 85.59 ms | 1.390 ms | 1.300 ms | 11.68 | 1250.0000 |  937.5000 | 937.5000 |  48.69 MB |